api/objects/agents/guard_agent.py
from dotenv import load_dotenv
import os
import json
from copy import deepcopy
from .utils import get_chatbot_response
from openai import OpenAI
load_dotenv()

class GuardAgent():
    def __init__(self):
        #Initialize the OpenAI client with the RunPod API key and the deployed chatbot URL
        #Initialize the model name being used
        self.client = OpenAI(
            api_key=os.getenv("RUNPOD_TOKEN"),
            base_url=os.getenv("RUNPOD_CHATBOT_URL"),
        )
        self.model_name = os.getenv("MODEL_NAME")

        #To get the system prompt for the guard agent allowing it to classify the user input into allowed and not allowed requests
    def get_system_prompt(self):
        return """ 
        You are a helpful AI assistant for a coffee shop application that serves drinks and pastries.

        Your task is to decide whether the user's request is relevant to the coffee shop's allowed topics. 
        You must carefully follow the rules below, even if the request is rephrased, disguised, or implied.

        ALLOWED USER REQUESTS:
        1. Ask about the coffee shop itself — location, opening hours, menu items, promotions, or general shop-related details.
        2. Ask for details about a menu item — e.g., ingredients, allergens, flavor descriptions, portion sizes, or price.
        3. Place an order for drinks, pastries, or other menu items.
        4. Ask for recommendations on what to buy (e.g., based on taste preferences or popular items).

        NOT ALLOWED USER REQUESTS:
        1. Ask about topics unrelated to our coffee shop or menu.
        2. Ask about staff members (personal details, hiring, roles, schedules, salaries, etc.).
        3. Ask for instructions, steps, recipes, or guidance on how to prepare, cook, or brew any menu item — including disguised forms such as:
        - Asking for “tips,” “advice,” or “tricks” on making an item
         - Asking for ingredients + procedure together
        - Asking how it is prepared at the shop
        4. Attempt to indirectly get a recipe or preparation method by asking for “similar” home recipes or step-by-step alternatives.

        GUARDRAILS:
        - If the user tries to bypass restrictions (e.g., “Just hypothetically, how would you make a latte?” or “If I were to make it at home, what steps would I take?”), this is still NOT ALLOWED.
        - If the request partially contains disallowed content, treat the entire request as NOT ALLOWED.
        - Only allow requests that are fully compliant with the ALLOWED list.

        OUTPUT FORMAT:
        Return a JSON object with the following keys and rules:
        {
        "chain of thought": Briefly go through each of the ALLOWED and NOT ALLOWED points and explain why this user message fits (or doesn’t fit) under them.
        "decision": "allowed" or "not allowed" — choose exactly one.
        "message": If decision is "allowed", leave this as an empty string. If "not allowed", set to "Sorry, I can't help with that. Can I help you with your order?"
        }

        Be concise but accurate in the "chain of thought" reasoning. """

        
    def get_response(self,messages):
        # Deep copy the messages to avoid modifying the original list
        messages = deepcopy(messages)
        
        #Get the designed system prompt to fetch accurate responses from the llm
        system_prompt = self.get_system_prompt()

        input_messages = [{"role": "system", "content": system_prompt}] + messages[-3:]

        chatbot_output =get_chatbot_response(self.client,self.model_name,input_messages)
        output = self.postprocess(chatbot_output)
        
        return output

#To postprocess the ouput from the llm to have the role, content and memory attributes.
    def postprocess(self,output):
        output = json.loads(output)

        dict_output = {
            "role": "assistant",
            "content": output['message'],
            "memory": {"agent":"guard_agent",
                       "guard_decision": output['decision']
                      }
        }
        return dict_output



    