api/objects/requirements.txt
pandas==2.2.3
openai==1.50.2
python-dotenv==1.0.1
pinecone==5.3.1
runpod==1.7.1